requirements.txt
# Web framework
fastapi==0.135.1
uvicorn[standard]==0.41.0

# Anthropic
anthropic==0.71.0

# Image rendering
Pillow>=11.0.0

# Gmail + GCP
google-api-python-client==2.189.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
google-cloud-firestore==2.25.0
google-cloud-storage==2.19.0

# Telegram bot
aiogram==3.15.0
redis==7.3.0

# Newsletter parsing
html2text>=2024.2.26
beautifulsoup4>=4.12.0
lxml>=5.0.0

# HTTP client
httpx==0.28.1

# Utilities
python-dotenv==1.0.0
